pyproject.toml
[project]
name = "home-ledger-api"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.95",
    "uvicorn[standard]>=0.22",
    "sqlalchemy>=1.4",
    "python-dotenv>=1.0",
    "httpx>=0.23",
    "pydantic-settings>=2.1",
    "psycopg2-binary>=2.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.23",
    "ruff>=0.4",
]
